CI: Only build dependencies not already available.
The build dependencies take a long time to build so only build them if
they don't already exist.

The approach taken is simple - compare the repository version of
"versions.txt" with a version installed at the time the cached versions
of the build dependencies were installed. If the versions differ at all
(only a simple string comparison is performed), rebuild that build
dependency.

This is a precursor to adding cache support to the Travis build.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>