CI: Enable caching of /usr/local for Travis.
Flagging /usr/local to Travis as a cache directory means that the
contents of this directory will be automatically restored on each build.
Hence, the first build to make use of this caching will populate the
cache (and this build will be slow), but subsequent builds can take
advantage of the pre-built dependencies for a faster turn-around.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>