Config: Allow kvm to be disabled for testing.
Clear Containers rely rely on kvm support, however this is not available
in all environments (such as Travis).

Add a new "--without-kvm-support" configure option to be specified which
will produce a "data/hypervisor.args" that is suitable for systems with
no kvm support. This is achieved by adding additional tags to
"data/hypervisor.args.in":

  - @HYPERVISOR_MACHINE_TYPE@
  - @HYPERVISOR_CPU_TYPE@
  - @HYPERVISOR_KVM_OPTIONS@

The values of these values is also shown in the configure summary under
the new "Hypervisor options" heading.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>