CI: Build and run tests inside docker.
A docker image provides a consistent environment to run the build and
tests inside.

Travis only provides Ubuntu Trusty meaning to support running in that
environment, it is necessary to build glib, json-glib, qemu-lite *and*
gcc6 (required by qemu-lite). This is possible, but the time taken to
build gcc6 and the amount of output generated exceed Travis job limits.

By using a docker image that already has the required library versions,
qemu-lite and the Clear Containers assets, the Travis build becomes a
lot simpler.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>